Moved follow/unfollow author to AuthorRepository.cs to follow SOLID
