+++ src/Chirp.Infrastructure/AuthorRepository.cs
@@ -1,29 +1,29 @@
-using System.Security.Principal;
-
 using Chirp.Core;
 
 using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Identity.UI.Services;
-using Microsoft.AspNetCore.Identity.UI.V4.Pages.Account.Internal;
-using Microsoft.Extensions.Logging;
+using Microsoft.EntityFrameworkCore;
 
 namespace Chirp.Infrastructure;
 
 public interface IAuthorRepository
 {
     Task<IdentityResult> AddAuthor(Author user, string? password = null);
+    Task FollowAuthor(string followerName, string authorName);
+    Task UnfollowAuthor(string followerName, string authorName);
 }
 
 public class AuthorRepository : IAuthorRepository
 {
     private readonly UserManager<Author> _userManager;
     private readonly IUserStore<Author> _userStore;
+    private readonly ChirpDBContext _dbContext;
 
 
-    public AuthorRepository (UserManager<Author> userManager, IUserStore<Author> userStore)
+    public AuthorRepository (UserManager<Author> userManager, IUserStore<Author> userStore, ChirpDBContext dbContext)
     {
         _userManager = userManager;
         _userStore = userStore;
+        _dbContext = dbContext;
     }
     
     public Task<IdentityResult> AddAuthor(Author user, string? password = null)
@@ -41,5 +41,44 @@ public class AuthorRepository : IAuthorRepository
             throw new NotSupportedException("The default UI requires a user store with email support.");
         }
         return (IUserEmailStore<Author>)_userStore;
+    }
+    
+    public async Task FollowAuthor(string followerName, string authorName)
+    {
+        var follower = await _dbContext.Authors
+            .Include(a => a.Following)
+            .FirstOrDefaultAsync(a => a.UserName == followerName);
+
+        var author = await _dbContext.Authors
+            .Include(a => a.Followers)
+            .FirstOrDefaultAsync(a => a.UserName == authorName);
+
+        if (follower == null || author == null)
+        {
+            throw new ArgumentException("User or author does not exist.");
+        }
+        
+        follower.Following.Add(author);
+
+        await _dbContext.SaveChangesAsync();
+    }
+    public async Task UnfollowAuthor(string followerName, string authorName)
+    {
+        var follower = await _dbContext.Authors
+            .Include(a => a.Following)
+            .FirstOrDefaultAsync(a => a.UserName == followerName);
+
+        var author = await _dbContext.Authors
+            .Include(a => a.Followers)
+            .FirstOrDefaultAsync(a => a.UserName == authorName);
+
+        if (follower == null || author == null)
+        {
+            throw new ArgumentException("User or author does not exist");
+        }
+        
+        follower.Following.Remove(author);
+
+        await _dbContext.SaveChangesAsync();
     }
 }
